starting columns and mock cards
