+++ client/src/App.jsx
@@ -1,5 +1,10 @@
 import React, { Component } from 'react'
+
+import Grid from 'material-ui/Grid'
+
 import { views as Kanban } from './Kanban'
+
+const styles = {}
 
 class App extends Component {
   constructor () {
@@ -9,13 +14,16 @@ class App extends Component {
   componentDidMoun () {
 
   }
-
   render () {
     return (
-      <main>
-        <Kanban />
-        <a href='/api/auth/google'><button>Login with Google</button></a>
-      </main>
+      <Grid container>
+        <div className='navBar' />
+        <Grid item xs={2} />
+        <Grid item xs={10}>
+          <Kanban />
+          <a href='/api/auth/google'><button>Login with Google</button></a>
+        </Grid>
+      </Grid>
     )
   }
 }
