+++ client/src/Kanban/views/Column.jsx
@@ -1,41 +1,13 @@
-import React, { Component } from 'react';
+import React, { Component } from 'react'
+import Card from './Card'
 
-class Form extends Component {
-  constructor(){
-    super();
-    this.state = {
-      message: ''
-    }
-
-    this.handleSubmit = this.handleSubmit.bind(this);
-    this.handleChange = this.handleChange.bind(this);
-  }
-
-  handleSubmit(event){
-    event.preventDefault();
-    console.log('Submit Event target value:', event.target.value);
-  }
-
-  handleChange(event){
-    console.log('Change event', event.target.value);
-    this.setState({message: event.target.value});
-
-  }
-
-  render(){
-    const { handleSubmit, handleChange } = this;
-    const { message } = this.state
-    return (
-      <div>
-        <form onSubmit={ handleSubmit }>
-          <input type="text" placeholder='Write your goals' autoFocus onChange={ handleChange }/>
-          <button>Submit</button>
-        </form>
-        <br />
-        <div>{ message }</div>
-      </div>
-    )
-  }
+const Column = ({header, goals}) => {
+  return (
+    <div>
+      <div>{header}</div>
+      <Card />
+    </div>
+  )
 }
 
-export default Form;
+export default Column
